Get rid of offlinemerger
We improved our merge strategy in the meantime, to the point where it
works perfectly for 99,9% of cases we get. We don't need an offline
merger anymore, and removing it will shave off plenty of CPU time.
